Leveraged pg-promise helpers to remove most of the manual string manipulation. Moved SQL generating code into its own module
