+++ db/connection-pool.ts
@@ -1,11 +1,9 @@
 import pgPromise from 'pg-promise';
 import configSet from '../config/config'
 
-const pgp = pgPromise({
+export const pgp = pgPromise({
 
 })
 
 const config = configSet.development
-const db = pgp(`postgres://${config.username}:${config.password}@${config.host}:${config.port}/${config.database}`);
-
-export default db;
+export const db = pgp(`postgres://${config.username}:${config.password}@${config.host}:${config.port}/${config.database}`);
